+++ src/metrics.rs
@@ -80,7 +80,8 @@ pub mod jet {
             Histogram, HistogramOpts, HistogramVec, IntCounter, IntCounterVec, IntGauge,
             IntGaugeVec, Opts,
         },
-        solana_sdk::{clock::Slot, pubkey::Pubkey},
+        solana_clock::Slot,
+        solana_pubkey::Pubkey,
         std::{
             net::SocketAddr,
             sync::{Mutex, Once},