+++ src/pubkey_challenger.rs
@@ -1,11 +1,10 @@
 use {
     core::fmt,
     serde::{Deserialize, Serialize},
-    solana_sdk::{
-        pubkey::Pubkey,
-        signature::{Keypair, Signature},
-        signer::Signer,
-    },
+    solana_keypair::Keypair,
+    solana_pubkey::Pubkey,
+    solana_signature::Signature,
+    solana_signer::Signer,
     std::{
         collections::HashMap,
         sync::Arc,
@@ -274,11 +273,7 @@ impl OneTimeAuthTokenStore for InMemoryTokenStore {
 
 #[cfg(test)]
 mod tests {
-    use {
-        super::*,
-        solana_sdk::signature::{Keypair, Signer},
-        std::time::Duration,
-    };
+    use {super::*, solana_keypair::Keypair, solana_signer::Signer, std::time::Duration};
 
     #[tokio::test]
     async fn test_sig_challenger() {