+++ src/util.rs
@@ -1,11 +1,10 @@
 use {
     futures::future::Either,
     serde::Deserialize,
-    solana_sdk::{
-        pubkey::Pubkey,
-        signature::{Keypair, Signature},
-        signer::Signer,
-    },
+    solana_keypair::Keypair,
+    solana_pubkey::Pubkey,
+    solana_signature::Signature,
+    solana_signer::{Signer, SignerError},
     std::{cmp::Ordering, future::Future, sync::Arc},
     tokio::{
         sync::{Mutex, oneshot, watch},
@@ -155,14 +154,11 @@ impl Signer for PubkeySigner {
         self.0.sign_message(message)
     }
 
-    fn try_pubkey(&self) -> Result<solana_sdk::pubkey::Pubkey, solana_sdk::signer::SignerError> {
+    fn try_pubkey(&self) -> Result<Pubkey, SignerError> {
         self.0.try_pubkey()
     }
 
-    fn try_sign_message(
-        &self,
-        message: &[u8],
-    ) -> Result<Signature, solana_sdk::signer::SignerError> {
+    fn try_sign_message(&self, message: &[u8]) -> Result<Signature, SignerError> {
         self.0.try_sign_message(message)
     }