+++ tests/test_transaction_retry_scheduler.rs
@@ -2,15 +2,13 @@ mod testkit;
 
 use {
     bytes::Bytes,
-    solana_sdk::{
-        hash::Hash,
-        message::{VersionedMessage, v0},
-        pubkey::Pubkey,
-        signature::Keypair,
-        signer::Signer,
-        system_instruction,
-        transaction::VersionedTransaction,
-    },
+    solana_hash::Hash,
+    solana_keypair::Keypair,
+    solana_message::{VersionedMessage, v0},
+    solana_program::system_instruction::transfer,
+    solana_pubkey::Pubkey,
+    solana_signer::Signer,
+    solana_transaction::versioned::VersionedTransaction,
     std::{
         sync::{Arc, RwLock as StdRwLock},
         time::Duration,
@@ -30,7 +28,7 @@ use {
 pub fn create_send_transaction_request(hash: Hash, max_resent: usize) -> SendTransactionRequest {
     let fake_wallet_keypair1 = Keypair::new();
     let fake_wallet_keypair2 = Keypair::new();
-    let instructions = vec![system_instruction::transfer(
+    let instructions = vec![transfer(
         &fake_wallet_keypair1.pubkey(),
         &fake_wallet_keypair2.pubkey(),
         10,